Selected tab icon changes color.
By converting the image into a QPixmap object it is possible to change its color. The image has to be transparent for the effect to work.
Code example:
(In this example we suppose that "i" is the QToolButton object)
```
color = QColor(255, 255, 255) # We create a QColor object with the RGB code of our choice.
pixmap = QPixmap(i.iconPath) # We generate a pixmap of the icon image.
painter = QPainter(pixmap) # We transform the pixmap into a QPainter object
painter.setCompositionMode(QPainter.CompositionMode_SourceIn)
painter.fillRect(pixmap.rect(), color) # We now fill the non alpha channel colors into the desired color.
painter.end() 
icon = QIcon(pixmap) # We convert the pixmap into a QIcon object
i.setIcon(icon) # We set the Icon to the QToolButton
```
The color change is compatible with dark mode (colors are slightly different).
This also means that dark mode icons are going to be #FFF (white) without requiring a second set of icon images.