improved server stability
fixed an issue where if the server had a slow CPU,  it could cause a crash since it wouldn't have the time to fully establish a connection with the database before running the client's request. Therefor, a verification to see if the connection to the database has been successfully established was added. This will improve the stability.